ArchiveRecord + impl moved into same Scala file; code cleanup. (#230)
